requirements.txt
flask~=3.1.1
opencv-python
numpy~=2.2.5
scikit-learn~=1.6.1
pillow~=11.2.1
torchvision~=0.22.0
open-clip-torch
flask-cors~=5.0.1
Werkzeug~=3.1.3
torch~=2.7.0
joblib~=1.5.0